requirements.txt
Django>=4.2.0,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
pandas>=1.5.0
openpyxl>=3.0.0
gunicorn>=20.1.0
whitenoise>=6.0.0
dj-database-url>=2.0.0
psycopg2-binary>=2.9.0
djoser==2.2.0
python-dotenv>=1.2.1
resend>=2.0.0
celery>=5.3.0
redis>=5.0.0